+++ apps/website-astro/src/ui/MenuBar/index.ts
@@ -1,0 +1,3 @@
+import MenuBar from './MenuBar.svelte';
+
+export { MenuBar };
