+++ apps/website-astro/src/ui/layouts/Layout.astro
@@ -15,7 +15,7 @@ import { type SupportedLanguages } from '@raulmelo/core/config';
 
 import DefaultSEO, { type Props as SeoProps } from '@/ui/DefaultSEO.astro';
 import Logo from '@/ui/Logo.astro';
-import MenuBar from '@/ui/MenuBar.svelte';
+import { MenuBar } from '@/ui/MenuBar';
 
 const { title, seo, lang } = Astro.props;
